getFirebaserc を切り出した
